[RENAME] 메뉴 예외 클래스들을 전역으로 수정 및 이동 #28
- domain.menu.exception → common.exception
- MenuExceptionHandler → GlobalExceptionHandler
